Add full Classic roster, pet and treasure pages, and Power+ unlock data
Scrape the complete Cookie Run Classic roster (96 cookies, 99 pets) from
cookierun.wiki with snake_case images, grades, release dates, and
localized skill/story text; add grid list pages with infinite scroll,
grade filters, and detail pages; make ids encode sort position so new
entries appear on top; add the E grade tier and remove SS; split Power+
into effect, requirement, and unlock goal fields; add admin-only creation
forms with image upload; fix htmx-boosted navigation clobbering list
pages; and add project documentation.

+++ config/config.v
@@ -5,7 +5,7 @@ import toml
 pub struct Config {
 pub:
 	host string = "127.0.0.1"
-	port u16 = 6785
+	port int = 6785
 	db_file string = "sqlite.db"
 }
